Injects ApiRequest.Options with accountId.
