+++ financial-connections/src/main/java/com/stripe/android/financialconnections/di/FinancialConnectionsSheetModule.kt
@@ -16,16 +16,16 @@ internal object FinancialConnectionsSheetModule {
     @Singleton
     @Provides
     fun providesFinancialConnectionsManifestRepository(
-        @Named(PUBLISHABLE_KEY) publishableKey: String,
         requestExecutor: FinancialConnectionsRequestExecutor,
         configuration: FinancialConnectionsSheet.Configuration,
         apiRequestFactory: ApiRequest.Factory,
+        apiOptions: ApiRequest.Options,
         logger: Logger
     ) = FinancialConnectionsManifestRepository(
-        publishableKey = publishableKey,
         requestExecutor = requestExecutor,
         configuration = configuration,
         apiRequestFactory = apiRequestFactory,
+        apiOptions = apiOptions,
         logger = logger,
         initialManifest = null
     )
